Adapt the shell generation to VIC

+++ vic/src/add_paths.rs
@@ -2,9 +2,52 @@ use crate::errors::Errcode;
 use crate::mounts::{create_directory, create_symlink, mount_directory};
 use nix::mount::MsFlags;
 use serde::{Deserialize, Serialize};
+use std::os::unix::prelude::PermissionsExt;
 use std::path::PathBuf;
 
-const BASE_MNT_FLAGS: [MsFlags; 2] = [MsFlags::MS_PRIVATE, MsFlags::MS_BIND];
+const BASE_MNT_FLAGS: [MsFlags; 3] = [MsFlags::MS_PRIVATE, MsFlags::MS_BIND, MsFlags::MS_REC];
+
+#[derive(Clone, Serialize, Deserialize)]
+#[serde(rename_all = "snake_case")]
+pub enum FsInitType {
+    Directory,
+    File { content: String },
+}
+
+pub type FsInit = std::collections::HashMap<PathBuf, FsInitEntry>;
+
+#[derive(Clone, Serialize, Deserialize)]
+pub struct FsInitEntry {
+    #[serde(rename = "type")]
+    fs_init_type: FsInitType,
+    perm: u32,
+}
+
+impl FsInitEntry {
+    pub fn create(&self, root: &PathBuf, dst: &PathBuf) -> Result<(), Errcode> {
+        let dst = if dst.starts_with("/") {
+            root.join(dst.strip_prefix("/").unwrap())
+        } else {
+            root.join(&dst)
+        };
+        match &self.fs_init_type {
+            FsInitType::Directory => std::fs::create_dir_all(&dst).unwrap(),
+            FsInitType::File { content } => {
+                if let Some(p) = dst.parent() {
+                    std::fs::create_dir_all(p).unwrap();
+                }
+                std::fs::write(&dst, content).unwrap();
+            }
+        }
+        let mode = u32::from_str_radix(format!("{}", self.perm).as_str(), 8).unwrap();
+        std::fs::set_permissions(&dst, std::fs::Permissions::from_mode(mode)).unwrap();
+        log::debug!(
+            "Created FS entry {dst:?} with permission {:?}",
+            std::fs::Permissions::from_mode(mode)
+        );
+        Ok(())
+    }
+}
 
 #[derive(Clone, Serialize, Deserialize)]
 #[serde(rename_all = "snake_case")]
@@ -23,9 +66,14 @@ impl Into<MsFlags> for &MountFlag {
 #[derive(Clone, Serialize, Deserialize)]
 #[serde(rename_all = "snake_case")]
 pub enum AddPathType {
-    Mount { flags: Vec<MountFlag> },
+    Mount {
+        flags: Vec<MountFlag>,
+        mount_type: Option<String>,
+    },
     Symlink,
-    SymlinkDirContent { exceptions: Vec<PathBuf> },
+    SymlinkDirContent {
+        exceptions: Vec<PathBuf>,
+    },
     Copy,
 }
 
@@ -48,13 +96,13 @@ impl AddPath {
     pub fn add_to_root(&self, root: &PathBuf) -> Result<(), Errcode> {
         let dst = self.get_root_dst(root);
         match &self.path_type {
-            AddPathType::Mount { flags } => {
+            AddPathType::Mount { flags, mount_type } => {
                 let mut mnt_flags = Vec::from(BASE_MNT_FLAGS);
                 for f in flags.iter() {
                     mnt_flags.push(f.into());
                 }
                 create_directory(&dst)?;
-                mount_directory(Some(&self.src), &dst, mnt_flags)?;
+                mount_directory(Some(&self.src), &dst, mount_type, mnt_flags)?;
             }
             AddPathType::Symlink => create_symlink(&self.src, &self.dst)?,
             AddPathType::SymlinkDirContent { exceptions } => {
@@ -73,7 +121,11 @@ impl AddPath {
                 }
             }
             AddPathType::Copy => {
-                std::fs::copy(&self.src, &dst).unwrap();
+                if !dst.exists() {
+                    std::fs::copy(&self.src, &dst).unwrap();
+                } else {
+                    log::info!("{dst:?} file exists, won't erase with a copy from host");
+                }
             }
         }
         Ok(())
@@ -82,9 +134,6 @@ impl AddPath {
     pub fn clean(&self, root: &PathBuf) -> Result<(), Errcode> {
         let dst = self.get_root_dst(root);
         match self.path_type {
-            AddPathType::Mount { .. } => {
-                // remove_empty_dir_tree(&dst)?;
-            }
             AddPathType::Symlink => {
                 if dst.is_symlink() {
                     std::fs::remove_file(dst).unwrap();
@@ -94,7 +143,7 @@ impl AddPath {
             }
             AddPathType::SymlinkDirContent { .. } => {
                 if !dst.is_dir() {
-                    log::warn!("Unable to clean symlinked-content dir {dst:?}, skipping ...");
+                    log::warn!("Unable to clean symlinked-content, {dst:?} is not a directory");
                     return Ok(());
                 }
                 for path in std::fs::read_dir(&dst).unwrap() {
@@ -104,13 +153,7 @@ impl AddPath {
                     }
                 }
             }
-            AddPathType::Copy => {
-                if !dst.is_file() {
-                    log::warn!("Unable to clean copied file {dst:?}, skipping ...");
-                } else {
-                    std::fs::remove_file(dst).unwrap();
-                }
-            }
+            _ => {}
         }
         Ok(())
     }
